Enforce stronger track semantics with the TrackBuilder (#48)
Consumers of track data (processors and target formats) should be able to assumed the data they get is valid. For example, that all note events are organized into groups with distinct start and end events. Maintaining this validity is complicated and it would be very easy for the author of a property to ignore edge-cases that can violate it. Example: The merge function can squash two notes of the same pitch so they overlap, confusing the group structure.

To enforce validity, tracks must now be constructed using a TrackBuilder. This checks validity as events are added and rejects events that would violate validity.

* Don't treat track info as a cache.

* First draft of TrackBuilder

* TrackBuilder: Handle several cases, but not zero length groups yet.

* Sketch algorithm that also handles zero-length groups

* Add a validate method to Track

* Add isTrackValid and some unit tests for the TrackBuilder.

* Fixes.

* The quantize function should use TrackBuilder instead of the sanitize function

* TrackBuilder fixes and more tests

* More testing

* Add the SimpleTrackBuilder

* Add a createEndEvent to TrackEvent

* TrackBuilder -> ValidTrackBuilder

* Add end events when groups unterminated

* Use a trackEventHolder to hold events

* Use the validTrackBuilder to simplify the excerptFunction

* Tests for excess duration and fix

* Merge function uses ValidTrackBuilder + test that shows why

* More functions using builders

* MonophonicSubtrackFunction uses builders

* Append uses a trackbuilder

* SmfParser uses a TrackBuilder

* Abandon SimpleTrackBuilder and commit to ValidTrackBuilder

* ValidTrackBuilder -> TrackBuilder

* Many tests using builder.

* Move TrackBuilder to Track folder

* Make Track::addEvent private to force TrackBuilder usage

* finishAndGetTrack can take a duration

* SImplify track duration handling

* Helpful asserts